Update README.md
The intend of this project is to help the Department of Foreign Literature from Zhejiang University to build a database based on their archives of journal PDFs in Chinese.
The goal is to extract thesis titles, author names, affiliated institutions, abstracts, keywords, journal titles, issue of the journal and page numbers. Then put them in an Excel sheet in respective columns.
If there is no such information then it will not be recorded.
This project is possible with the help and advices of Yuqing Ye and Yifei Li @ Zhejiang University.